feat(home): add the flood risk demo and simplify how it works
The homepage now runs two recordings. A new one leads the `#demo` section
under the hero. It shows an agent that answers questions about flood risk in
Moldova from published Portolan data. The catalog build recording moves down
into "How it works", where it stands as evidence for the four steps.

The `#demo` section gains a heading and one sentence. Before this it held a
video and a caption. A dark terminal poster cannot say what the run proves, so
the section says it.

"How it works" loses `PipelineFigure`. The section explained one pipeline
three times. The video showed it, the figure drew it, and the steps named it.
The steps now run as a numbered spine on one rule instead of four cards.
