+++ messages/ar.json
@@ -86,33 +86,34 @@
     "eyebrow": "كيف يعمل",
     "title": "كيف يعمل",
     "intro": "استخدم <cli>واجهة سطر أوامر Portolan</cli> أو <skills>مهارات الوكيل</skills> لتحويل بياناتك الحالية إلى كتالوج Portolan.",
-    "figAlt": "تتدفق الملفات والخدمات الحالية إلى portolan-cli، الذي يدفع كتالوجًا إلى تخزين الكائنات. وتقرأه QGIS وPython والوكلاء مباشرة.",
     "steps": {
       "convert": {
         "id": "01",
         "title": "التحويل",
-        "description": "تُحوَّل بياناتك الحالية إلى صيغ مصمّمة للسحابة. سواء عملت بملفات Shapefile وGeoPackage أو بنقطة WFS أو بخادم Esri، ستحصل على <m>GeoParquet</m> للبيانات الشعاعية و<m>COG</m> للبيانات النقطية."
+        "description": "سواء عملت بملفات Shapefile وGeoPackage أو بنقطة WFS أو بخادم Esri، تصبح بياناتك <m>GeoParquet</m> للبيانات الشعاعية و<m>COG</m> للبيانات النقطية."
       },
       "catalog": {
         "id": "02",
         "title": "الفهرسة",
-        "description": "تُنظَّم بياناتك المهيأة للسحابة ككتالوج <m>STAC</m> مع بيانات وصفية وتوثيق واضحين، بما في ذلك ملفات README و<m>AGENTS.md</m>. وهذا يجعل الكتالوج سهل العثور عليه واستكشافه واستخدامه للناس والوكلاء معًا."
+        "description": "تُنظَّم بياناتك المهيأة للسحابة ككتالوج <m>STAC</m> مع بيانات وصفية وتوثيق واضحين، وهذا يجعل الكتالوج سهل العثور عليه وفهمه واستخدامه للناس والوكلاء معًا."
       },
       "publish": {
         "id": "03",
         "title": "النشر",
-        "description": "يُنشر كتالوجك على أي تخزين متوافق مع S3 مثل AWS أو GCS أو Hetzner أو MinIO. لا توجد خوادم ولا واجهات برمجية تحتاج إعدادًا. ومشاركته تعني إرسال الرابط إلى شخص ما."
+        "description": "يُنشر كتالوجك على أي تخزين متوافق مع S3 مثل AWS أو GCS أو Hetzner أو MinIO، دون خوادم أو واجهات برمجية تحتاج إعدادًا."
       },
       "use": {
         "id": "04",
         "title": "الاستخدام",
-        "description": "تستطيع الآن استخدام البيانات مباشرة في أدوات مثل QGIS أو ArcGIS أو Python، أو تمرير رابط الكتالوج إلى وكيل ليبدأ العمل عبر مهارة الاستهلاك في Portolan."
+        "description": "تستطيع الآن نقل بياناتك مباشرة إلى أدوات مثل QGIS أو ArcGIS أو Python، أو تمرير رابط الكتالوج إلى وكيل ليبدأ العمل."
       }
     }
   },
   "demo": {
     "navLabel": "عرض توضيحي",
-    "caption": "إنشاء <catalog>فهرس Portolan</catalog> من البداية إلى النهاية من بيانات الإسكان في فيلادلفيا.",
+    "title": "طريقة أفضل للعمل مع البيانات الجغرافية المكانية",
+    "blurb": "شاهد وكيلا يجيب عن أسئلة بلغة طبيعية حول مخاطر الفيضانات في مولدوفا باستخدام بيانات من <registry>سجل Portolan</registry>.",
+    "captionPhiladelphia": "إنشاء <catalog>فهرس Portolan</catalog> من البداية إلى النهاية من بيانات الإسكان في فيلادلفيا.",
     "play": "شغل العرض",
     "pause": "أوقف العرض مؤقتا",
     "fullscreen": "اعرض العرض بملء الشاشة",
