+++ src/components/home-page.tsx
@@ -5,14 +5,19 @@ import { Link } from "@/i18n/navigation";
 import { GlyphMap } from "./glyph-map";
 import { SiteShell } from "./site-rail";
 import { DemoSection } from "./demo-section";
+import { VideoFigure } from "./video-figure";
 import { EcosystemSection } from "./ecosystem-section";
 import { InvolvedSection } from "./involved-section";
 import { GetInvolvedSection } from "./get-involved-section";
 import { WhoForSection } from "./who-for-section";
 import { CoverageSection } from "./coverage-section";
-import { PipelineFigure } from "./pipeline-figure";
 import { Btn, DirArrow, monoChunk } from "./ui";
 import type { Catalog } from "@/lib/catalogs";
+
+/** Aspect ratio of the encoded catalog-build recording, 1454x1118. It
+ *  reserves the box before metadata loads, so the poster does not shift the
+ *  page when it paints. */
+const DEMO_ONE_RATIO = "1454 / 1118";
 
 interface HomePageProps {
   catalogs?: Catalog[];
@@ -88,7 +93,9 @@ export function HomePage({ catalogs = [] }: HomePageProps) {
       </section>
 
       {/* The recorded demo. It answers the hero before the page explains
-          itself: one run from a source portal to a published catalog. */}
+          itself: one run that reads published Portolan data and measures
+          flood risk with it. The companion run, which builds the catalog,
+          plays inside "How it works" below. */}
       <DemoSection />
 
       {/* Why Portolan */}
@@ -167,41 +174,62 @@ export function HomePage({ catalogs = [] }: HomePageProps) {
               })}
             </p>
           </div>
-          <div className="mb-[var(--p-pad-lg)]">
-            <PipelineFigure />
-          </div>
-          {/* Stepper strip: four cells in one near-black frame, soft interior
-              dividers (wrap-aware for the 1/2/4-col responsive steps), faint
-              tint on hover. Number badge (mono, reversed) + title. The badges
-              already number the sequence, so the cells carry no arrow. An
-              arrow here read as a link and said nothing the order did not. */}
-          <div className="border border-p-line grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4">
-            {howSteps.map((step) => (
-              <div
-                key={step}
-                className="group p-6 border-p-line transition-colors duration-300
-                  border-b [&:last-child]:border-b-0
-                  sm:[&:nth-child(2n)]:border-s
-                  sm:[&:nth-child(n+3)]:border-b-0
-                  lg:border-b-0
-                  lg:[&:not(:first-child)]:border-s
-                  hover:bg-[color-mix(in_srgb,var(--p-primary)_4%,var(--p-paper))]"
-              >
-                <div className="flex items-center gap-2.5">
-                  <span className="font-mono text-eyebrow text-p-bg bg-p-primary px-1.5 py-0.5 leading-none tracking-[0.04em] rtl:tracking-normal">
+          {/* Two columns, mirroring "Explore catalogs" below: the media takes
+              the wide side and the list takes the narrow one, with the sides
+              swapped so the two sections do not stack the same way twice.
+
+              The steps name the pipeline and the run shows it. Side by side,
+              the reader reads a step and sees it happen without scrolling
+              between the two. */}
+          <div className="mt-10 grid grid-cols-1 items-center gap-6 xl:grid-cols-[minmax(0,55fr)_minmax(0,45fr)] xl:gap-8">
+            {/* The run leads on the wide column, because it carries the
+                section. */}
+            <VideoFigure
+              src="/video/portolan-demo-one.mp4"
+              poster="/video/portolan-demo-one.jpg"
+              ratio={DEMO_ONE_RATIO}
+              caption={t.rich("demo.captionPhiladelphia", {
+                catalog: (chunks) => (
+                  <a
+                    href="https://source.coop/nlebovits/phl-housing-demo"
+                    className="text-p-primary underline underline-offset-2 transition-colors hover:text-p-ink"
+                  >
+                    {chunks}
+                  </a>
+                ),
+              })}
+            />
+
+            {/* One spine, four stops. The steps were four bordered cards,
+                which put a third card grid on a page that already runs two,
+                and made a sequence read as four peer objects. A single rule
+                carries the order instead: the numbers sit on the line and the
+                text hangs off it. No box, no offset, and no hover state,
+                because none of this is clickable. The numbers stay because
+                the pipeline runs in this order and the reader needs it.
+
+                The rule is the near-black structural ink, not the soft
+                interior tier, because nothing encloses this list. */}
+            <ol className="relative ms-[13px] border-s border-p-line">
+              {howSteps.map((step) => (
+                <li key={step} className="relative ps-7 pb-9 last:pb-0 sm:ps-8">
+                  {/* Pulled back over the rule by half its own width, with the
+                      page ground behind it, so the line reads as broken by the
+                      number rather than crossed out by it. */}
+                  <span className="absolute start-0 top-[4px] -ms-[13px] bg-p-bg px-1.5 font-mono text-eyebrow leading-none text-p-primary">
                     {t(`howItWorks.steps.${step}.id`)}
                   </span>
-                  <h3 className="text-card-title font-bold tracking-[-0.02em]">
+                  <h3 className="text-card-title font-bold tracking-[-0.02em] leading-none">
                     {t(`howItWorks.steps.${step}.title`)}
                   </h3>
-                </div>
-                <p className="text-body text-p-ink-2 leading-relaxed text-pretty mt-3">
-                  {t.rich(`howItWorks.steps.${step}.description`, {
-                    m: monoChunk,
-                  })}
-                </p>
-              </div>
-            ))}
+                  <p className="text-body text-p-ink-2 leading-relaxed text-pretty mt-3">
+                    {t.rich(`howItWorks.steps.${step}.description`, {
+                      m: monoChunk,
+                    })}
+                  </p>
+                </li>
+              ))}
+            </ol>
           </div>
         </div>
       </section>
