+++ src/components/index.ts
@@ -3,6 +3,7 @@ export { GlyphMap } from "./glyph-map";
 export { PageHero } from "./page-hero";
 export { LiveCount } from "./live-count";
 export { DemoSection } from "./demo-section";
+export { VideoFigure } from "./video-figure";
 export { HomePage } from "./home-page";
 export { FaqPage } from "./faq-page";
 export { TalksPage } from "./talks-page";
